requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
discord.py==2.3.1
flask==3.1.2
frozenlist==1.8.0
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
markdownify==1.2.0
markupsafe==3.0.3
multidict==6.7.0
propcache==0.4.1
python-dotenv==1.0.0
six==1.17.0
soupsieve==2.8
typing-extensions==4.15.0
urllib3==1.26.12
werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
